* Implement the download algorithm as explained in the wiki : http://dev.aybabtu.org/projects/pmp/wiki/Protocol_core-core#Downloading-thread ** Choose the best chunk to download : choose randomly among the chunks which have the fastest free peer. * Reduce the file buffer size

+++ application/Core/FileManager/IDataReader.h
@@ -36,7 +36,7 @@ namespace FM
         * @exception ChunkDeletedException
         * @exception ChunkNotCompletedException
         */
-      virtual quint64 read(char* buffer, uint offset) = 0;
+      virtual int read(char* buffer, uint offset) = 0;
    };
 }
